% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_blue.R, R/theme_purple.R
\name{theme_blue}
\alias{theme_blue}
\alias{theme_purple}
\title{Best color themes for ggplot2}
\usage{
theme_blue()

theme_purple()
}
\arguments{
\item{...}{arguments passed to function}
}
\description{
These are complete themes which control all non-data display.
}
